+++ Plugins/FreeRDP/Server/ParameterFreeRDPServer.cpp
@@ -1,0 +1,163 @@
+// Author: Kang Lin <kl222@126.com>
+
+#include "ParameterFreeRDPServer.h"
+
+CParameterFreeRDPServer::CParameterFreeRDPServer(
+    QObject *parent, const QString &szPrefix)
+    : CParameterServer{parent, szPrefix}
+    , m_bTlsSecurity(true)
+    , m_bRdpSecurity(true)
+    , m_bNlaSecurity(false)
+    , m_bNlaExtSecurity(false)
+#ifdef WITH_SHADOW_X11
+    , m_bAuthentication(true)
+#else
+    , m_bAuthentication(false)
+#endif
+    , m_MayView(true)
+    , m_MayInteract(true)
+{
+    m_Net.SetPort(3389);
+}
+
+int CParameterFreeRDPServer::OnLoad(QSettings &set)
+{
+    int nRet = 0;
+    nRet = CParameterServer::OnLoad(set);
+    if(nRet) return nRet;
+
+    setTlsSecurity(set.value("Security/Tls", getTlsSecurity()).toBool());
+    setRdpSecurity(set.value("Security/Rdp", getRdpSecurity()).toBool());
+    setNlaSecurity(set.value("Security/Nla", getNlaSecurity()).toBool());
+    setNlaExtSecurity(set.value("Security/NlaExt", getNlaExtSecurity()).toBool());
+    setSamFile(set.value("Security/Nla/SamFIle", getSamFile()).toString());
+
+    setAuthentication(set.value("Server/Authentication", getAuthentication()).toBool());
+    setMayView(set.value("Server/MayView", getMayView()).toBool());
+    setMayInteract(set.value("Server/MayInteract", getMayInteract()).toBool());
+
+    return nRet;
+}
+
+int CParameterFreeRDPServer::OnSave(QSettings &set)
+{
+    int nRet = 0;
+    nRet = CParameterServer::OnSave(set);
+    if(nRet) return nRet;
+
+    set.setValue("Security/Tls", getTlsSecurity());
+    set.setValue("Security/Rdp", getRdpSecurity());
+    set.setValue("Security/Nla", getNlaSecurity());
+    set.setValue("Security/NlaExt", getNlaExtSecurity());
+    set.setValue("Security/Nla/SamFIle", getSamFile());
+
+    set.setValue("Server/Authentication", getAuthentication());
+    set.setValue("Server/MayView", getMayView());
+    set.setValue("Server/MayInteract", getMayInteract());
+
+    return nRet;
+}
+
+bool CParameterFreeRDPServer::getTlsSecurity() const
+{
+    return m_bTlsSecurity;
+}
+
+void CParameterFreeRDPServer::setTlsSecurity(bool newTlsSecurity)
+{
+    if (m_bTlsSecurity == newTlsSecurity)
+        return;
+    m_bTlsSecurity = newTlsSecurity;
+    emit sigTlsSecurityChanged();
+}
+
+bool CParameterFreeRDPServer::getRdpSecurity() const
+{
+    return m_bRdpSecurity;
+}
+
+void CParameterFreeRDPServer::setRdpSecurity(bool newRdpSecurity)
+{
+    if (m_bRdpSecurity == newRdpSecurity)
+        return;
+    m_bRdpSecurity = newRdpSecurity;
+    emit sigRdpSecurityChanged();
+}
+
+bool CParameterFreeRDPServer::getNlaSecurity() const
+{
+    return m_bNlaSecurity;
+}
+
+void CParameterFreeRDPServer::setNlaSecurity(bool newNlaSecurity)
+{
+    if (m_bNlaSecurity == newNlaSecurity)
+        return;
+    m_bNlaSecurity = newNlaSecurity;
+    emit sigNlaecurityChanged();
+}
+
+bool CParameterFreeRDPServer::getNlaExtSecurity() const
+{
+    return m_bNlaExtSecurity;
+}
+
+void CParameterFreeRDPServer::setNlaExtSecurity(bool newNlaExtSecurity)
+{
+    if (m_bNlaExtSecurity == newNlaExtSecurity)
+        return;
+    m_bNlaExtSecurity = newNlaExtSecurity;
+    emit sigNlaExtSecurityChanged();
+}
+
+bool CParameterFreeRDPServer::getAuthentication() const
+{
+    return m_bAuthentication;
+}
+
+void CParameterFreeRDPServer::setAuthentication(bool newAuthentication)
+{
+    if (m_bAuthentication == newAuthentication)
+        return;
+    m_bAuthentication = newAuthentication;
+    emit sigAuthenticationChanged();
+}
+
+bool CParameterFreeRDPServer::getMayView() const
+{
+    return m_MayView;
+}
+
+void CParameterFreeRDPServer::setMayView(bool newMayView)
+{
+    if (m_MayView == newMayView)
+        return;
+    m_MayView = newMayView;
+    emit sigMayViewChanged();
+}
+
+bool CParameterFreeRDPServer::getMayInteract() const
+{
+    return m_MayInteract;
+}
+
+void CParameterFreeRDPServer::setMayInteract(bool newMayInteract)
+{
+    if (m_MayInteract == newMayInteract)
+        return;
+    m_MayInteract = newMayInteract;
+    emit sigMayInteractChanged();
+}
+
+const QString &CParameterFreeRDPServer::getSamFile() const
+{
+    return m_szSamFile;
+}
+
+void CParameterFreeRDPServer::setSamFile(const QString &newSamFile)
+{
+    if (m_szSamFile == newSamFile)
+        return;
+    m_szSamFile = newSamFile;
+    emit sigSamFileChanged();
+}