+++ Plugins/FreeRDP/Server/Resource/Translations/FreeRDPServer_zh_TW.ts
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="zh_TW">
+<context>
+    <name>CPluginFreeRDPServer</name>
+    <message>
+        <location filename="../../PluginFreeRDPServer.cpp" line="56"/>
+        <source>Dependency libraries</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../PluginFreeRDPServer.cpp" line="57"/>
+        <source>FreeRDP</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../PluginFreeRDPServer.cpp" line="58"/>
+        <source>version: </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../PluginFreeRDPServer.cpp" line="61"/>
+        <source>Build version: </source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../PluginFreeRDPServer.cpp" line="67"/>
+        <source>Build date: </source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+</TS>